requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-core==0.3.0
langchain-ollama==0.2.0 # For embeddings
langchain-unstructured==0.1.4 # Document loading

